we're almost there... i can almost taste it
the most updated version of the game.  we still need to polish a lot but most things are implemented.  Currently the dialogue before each level isn't set up and the health bars for enemies and the boss aren't either.  Timing for animations could be better and there could be a few sounds that aren't added in yet.